Optionally parse Kafka message content as JSON (#15)
Use a new "parseAsJSON" parameter to decide to attempt to parse the message value as JSON. It is expected that this parameter will be set by the feed action, which can then be set by the user when creating the trigger.

There may be some existing stored triggers that do no have the "isJSONData" field set. Make sure these are restored properly as well.