Implement "The Doctor"
The Doctor is a separate thread that monitors the health of consumers.  If it notices that a consumer is sick (should be running, but isn't) it will attempt to restart that consumer.

This required making a variety of classes and data structures thread safe, as they may be accessed by the Doctor and the web app simultaneously. "currentState" reflects how the consumer believes it is running, while "desiredState" is the one the consumer should actively work toward transitioning to. "desiredState" can only be set to Restart or Dead.

This also required implementing explicit states in Consumers that allow them to be restarted gracefully. The biggest change here is that the Consumer no longer *is* a Thread, but instead has a field which is the ConsumerThread.