Automatically restart consumers that haven't polled in a while
There seems to be an issue where the kafka-python client gets into an error-handling loop. This causes poll() to not succeed, but also not throw an exception either.

This code adds logic to the Doctor to detect when a Consumer has not successfully polled in a while (2 seconds) and automatically restarts it.

Thanks @dubeejw