Actively monitor consumers - attempt to heal sick ones individually (#54)
* Implement "The Doctor"

The Doctor is a separate thread that monitors the health of consumers.  If it notices that a consumer is sick (should be running, but isn't) it will attempt to restart that consumer.

This required making a variety of classes and data structures thread safe, as they may be accessed by the Doctor and the web app simultaneously. "currentState" reflects how the consumer believes it is running, while "desiredState" is the one the consumer should actively work toward transitioning to. "desiredState" can only be set to Restart or Dead.

This also required implementing explicit states in Consumers that allow them to be restarted gracefully. The biggest change here is that the Consumer no longer *is* a Thread, but instead has a field which is the ConsumerThread.

* Adjust logging

Change logging levels on a variety of messages - mostly changing info to debug

* Keep track of how many times a Consumer is restarted

This is more for informational purposes, and isn't necessarily actionable in an automated way.

* Automatically restart consumers that haven't polled in a while

There seems to be an issue where the kafka-python client gets into an error-handling loop. This causes poll() to not succeed, but also not throw an exception either.

This code adds logic to the Doctor to detect when a Consumer has not successfully polled in a while (2 seconds) and automatically restarts it.

Thanks @dubeejw