Automatically restart the container if the health endpoint fails
Use the docker HEALTHCHECK instruction to curl the /health endpoint every five minutes.
If the endpoint does not respond within 30 seconds, cause the container's restart policy to kick in by killing the main python process.
Unfortunately, docker 1.12 does not (yet) support automatically restarting containers in response to healthcheck failures. This is a stop-gap measure to get the container to restart.