produce actions should handle binary keys and values
The idea is that, in order to use this action with binary data for message keys and values, the action invoker must first base64 encode the binary data. After all, this binary data must pass through a REST call in order to get to the produce action in the first place. However, in order to get the actual binary data produced in their message, the action must first decode that data back into the original bytes. To do that, the action invoker must use the isBinaryKey and/or isBinaryValue parameters set to true, as needed.