Reuse the basic health test trigger, if it already exists
Check if the health test trigger already exists, if it does, reuse it. If not, create it and do not delete it at the end of the test.

Added some additional logic to prevent cross-pollination of triggers being fired by other tests simultaneously running and producing messages to the same topic. Previously, this was prevented by having a trigger name that contained a unique timestamp. Now we poll for up to 5 activations that have occurred since just before the message was produced. Existing code already handled multiple activations, and rooted through them all to find the one that contains the expected message.