Account for the fact that the test Kafka instance is shared
Because the test kafka instance/topic are shared, it is possible that the number of trigger activations includes messages that were produced by entities other than the currently-running test. Account for this by polling for a larger-than-expected number of trigger activations and then sifting through those to ensure they match expectations.